This really seems to work...
Well at least it seems to. There are some bugs i still have to fix,
but overall it looks pretty good so far. The mtsp modules trigger, when they
should, i think my laptop can speak drunken russian now, and the software
wants to die, but overall it seems to be fine. Thank god this works the way i
thought it would.