Cleanup in game and sound for sequences
Implemented sound for sequence triggers so that an update will now trigger a
sound. I still have to implement sound error handling for this, as it is now
possible to kill alsa via just playing too much sound at once, hence forcing
the oom killer to forcefully shutdown alsa. yes, i know this sounds stupid, but
the oom killer knows what it does, so i am currently not able to help it

Signed-off-by: Tyrolyean <tyrolyean@tyrolyean.net>