Follow up to the emergency commit
Today I was called in because "it was like hell on earth". The wizard wasn't
working and PANIC!!!!!! I was uploading a new version and today it wa broken.
I thought maybe I had done a mistake and rolled back to the old version. But
nope, it wasn't really working. I then made the previous emergency commit, but
it still wasn't really working, though a bit better. A qick look at the sent
and received data showed, that the devices sent data, but a bit or two were
always wrong. I haven't had time yet to confirm this, but I suspect the
degradation of the materials used in the ESD-Team PCBs make communication way
more difficult at the almost 1/2 a Megabaud. This is really concerning, as the
PCBs are only one year old, though we have no idea where they did what
before they came to us...

Signed-off-by: Tyrolyean <tyrolyean@tyrolyean.net>