Version 7
This version includes support for progress in dependencies, adds a bit more
clarity to the code I wrote earlier, fixes a few bugs, and so on...

For progress in dependencies, a floating point number is used, the original
scheme by which a dependency is either forefilled or not is STILL in use to
validate events. This is only an addition to the already existing code, not a
supplement!

A few bug fixes were introduced to make my life easier... Fuck...

Signed-off-by: tyrolyean <tyrolyean@gmx.at>